+++ src/Infrastructure/Laravel/config/config.php
@@ -76,7 +76,7 @@ return [
      * - unit_based: derive excluding amount per unit and multiply by quantity.
      * - line_based: derive excluding amount from the full line including amount.
      */
-    'vat_rounding_strategy' => 'unit_based',
+    'vat_rounding_strategy' => \Thinktomorrow\Trader\Domain\Common\Vat\VatRoundingStrategy::getDefault()->value,
 
     /**
      * If this is true, the shop allows vat exemption for international business shoppers.